week2 homework
