+++ src/App.jsx
@@ -1,38 +1,17 @@
 import { useState } from 'react'
-import './App.css'
-
-// Array
-const todoList = [
-  { 
-    id: 1,
-    title: 'Homework' 
-  },
-  { 
-    id: 2,
-    title: 'Buy groceries' 
-  },
-  { id: 3,
-    title: 'Make dinner'
-  },
-];
+import './App.css';
+import TodoList from './TodoList';
+import AddTodoForm from './AddTodoForm';
 
 function App() {
 
   return(
   <div>
     <h1>TODO LIST</h1>
-    <ul> 
-         {todoList.map( function(item) {
-            return (
-            <li key={item.id}> 
-              {item.title} 
-            </li>
-            );
-         })}
-     </ul>
+    <AddTodoForm />
+    <TodoList />
   </div>
   );
 }
-//console.log(App());
 
 export default App;
